Implemented timestamp on blocks and transaction through approximation.
