+++ src/constants.ts
@@ -14,9 +14,8 @@ export const IDB_BLOCKS_CACHE = "zen_exp_blocks";
 
 /** BLOCK TIME */
 
-export const BLOCK_TIME_REFERENCE = {bNum: 1550361, timestamp: 1681286820};
-export const ODIN_BLOCK_TIME_REFERENCE = {bNum: 904587, timestamp: 1681287060};
-export const BLOCK_TIME = 120;
+export const LOKI_BLOCK_TIME_REFERENCE = {bNum: 1550361, timestamp: 1681294680}; // LOKI 1550361 at 10:07
+export const ODIN_BLOCK_TIME_REFERENCE = {bNum: 904587, timestamp: 1681294800}; // ODIN 904587 at 10:11 
 
-// ODIN 904587 at 10:11 
-// LOKI 1550361 at 10:07
+export const BLOCK_TIME = 120;
+export const BLOCK_TIME_PRE_REF= 60;
